🏗️ macros: ReplyError derive no longer requires specific field order
Now that we require `alloc` in core, we can simplify the generated
`Deserialize` impl by making use of `serde_derive`, which means we no
longer need to rely on the fields of a reply error message to be in a
specific order.

However, this unfortunately also means that we no longer have the
workaround a serde bug[1] that fails the deserialization of error messages
with empty parameter list. Fortunately, since it makes no sense to encode
empty paramater list, this will no longer be a real issue after the only
known impl to encode empty parameter list, soon stops doing so[2].

[1]: https://github.com/serde-rs/serde/issues/2045#issuecomment-3148593639
[2]: https://github.com/systemd/systemd/pull/38922