♻️  core: Simplify the server code
By keeping the connection halves together in one Vec. I distinctively
remember that I had a good reason to keep the 2 halves seperate but I
can't remember what it was, can't figure out why it would be needed and
all the tests are passing so I'm assuming for now that, this change is
safe and good.